% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{calculateTL}
\alias{calculateTL}
\title{Utility function to calculate training loads}
\usage{
calculateTL(df)
}
\arguments{
\item{df}{data frame with Date and load columns}
}
\value{
data frame with Date, ATL, CTL, TSS columns
}
\description{
The function calculates Acute Training Load (ATL), Chronic Training
Load (CTL), and Training Stress Score (TSS) based on the provided load data.
}
