% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeatmap.R
\name{calendarHeatmap}
\alias{calendarHeatmap}
\title{Calendar Heatmap}
\usage{
calendarHeatmap(dates, values, title = "", subtitle = "", legendtitle = "")
}
\arguments{
\item{dates}{A vector containing the dates in \code{Date} format.}

\item{values}{A vector containing the corresponding values as numeric.}

\item{title}{Main plot title (optional).}

\item{subtitle}{Main plot subtitle (optional).}

\item{legendtitle}{Legend title (optional).}
}
\value{
ggplot object
}
\description{
Creates a colour-coded calendar to visualise time series data
}
