% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeatmap.R
\name{calendarView}
\alias{calendarView}
\title{Calendar View}
\usage{
calendarView(df_all, from, to)
}
\arguments{
\item{df_all}{Data frame with all activity data}

\item{from}{Start date in "YYYY-MM-DD" format}

\item{to}{End date in "YYYY-MM-DD" format}
}
\value{
Saves calendar heatmaps as single PNG file
}
\description{
Creates calendar heatmaps for number of runs and distance per day
}
