% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndCalc.R
\name{compare2target}
\alias{compare2target}
\title{Function to filter activities and calculate cumulative distance}
\usage{
compare2target(activity, fromStr, toStr, df)
}
\arguments{
\item{activity}{character string of activity type to use as a filter}

\item{fromStr}{character string of form "YYYY-MM-DD"}

\item{toStr}{character string of form "YYYY-MM-DD"}

\item{df}{data frame with all activities}
}
\value{
data frame with activities of type activity within date window
}
\description{
Function to filter activities and calculate cumulative distance
}
