% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_years.R
\name{compare_years}
\alias{compare_years}
\title{Comparison of yearly Garmin Data}
\usage{
compare_years(
  activity = "Running",
  datadir = "Data",
  from = "",
  to = "",
  minimal = FALSE
)
}
\arguments{
\item{activity}{character string of activity type to use as a filter}

\item{datadir}{character string of folder name where CSV files are stored}

\item{from}{optional character string in "YYYY-MM-DD" format to filter start date}

\item{to}{optional character string in "YYYY-MM-DD" format to filter end date}

\item{minimal}{logical indicating whether to load minimal columns (default FALSE)}
}
\value{
ggplots comparing years
}
\description{
Comparison of yearly Garmin Data
}
