% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{find_form}
\alias{find_form}
\title{Wrapper function to find and plot Form (CTL, ATL, TSS)}
\usage{
find_form(from, to)
}
\arguments{
\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}
}
\value{
saves plot to Output/Plots folder
}
\description{
Wrapper function to find and plot Form (CTL, ATL, TSS)
}
