% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressToTarget.R
\name{get_data_subset}
\alias{get_data_subset}
\title{Function to subset activities within a window and find cumulative distance}
\usage{
get_data_subset(df, from, to)
}
\arguments{
\item{df}{data frame with all activities}

\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}
}
\value{
data frame with activities of type activity within date window
}
\description{
Function to subset activities within a window and find cumulative distance
}
