% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{getWindowActivities}
\alias{getWindowActivities}
\title{Utility function to get activities of a certain type within a date window}
\usage{
getWindowActivities(activity, fromStr, toStr, df)
}
\arguments{
\item{activity}{string of activity type to use as a filter}

\item{fromStr}{string of form "YYYY-MM-DD"}

\item{toStr}{string of form "YYYY-MM-DD"}

\item{df}{data frame with all activities}
}
\value{
data frame with activities of type activity within date window
}
\description{
Utility function to get activities of a certain type within a date window
}
\keyword{internal}
