% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressToTarget.R
\name{make_linear_target}
\alias{make_linear_target}
\title{Function to assemble target cumulative distance data frame}
\usage{
make_linear_target(from, to, km)
}
\arguments{
\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}

\item{km}{numeric target distance in km for the period from from to to}
}
\value{
data frame with Date, Cumulative.Distance columns
}
\description{
Function to assemble target cumulative distance data frame
}
\keyword{internal}
