% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{makeDateDF}
\alias{makeDateDF}
\title{Utility function to make date data frame}
\usage{
makeDateDF(fromStr, toStr)
}
\arguments{
\item{fromStr}{string of form "YYYY-MM-DD"}

\item{toStr}{string of form "YYYY-MM-DD"}
}
\value{
data frame with Date, ATL, CTL columns
}
\description{
Utility function to make date data frame
}
\keyword{internal}
