% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndCalc.R
\name{maketarget}
\alias{maketarget}
\title{Function to assemble target cumulative distance data frame}
\usage{
maketarget(fromStr, toStr, km)
}
\arguments{
\item{fromStr}{character string of form "YYYY-MM-DD"}

\item{toStr}{character string of form "YYYY-MM-DD"}

\item{km}{numeric target distance in km for the period from fromStr to toStr}
}
\value{
data frame with Date, Cumulative.Distance columns
}
\description{
Function to assemble target cumulative distance data frame
}
\keyword{internal}
