% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessGarminData.R
\name{preprocess_garmin_data}
\alias{preprocess_garmin_data}
\title{Preprocess Garmin activity data}
\usage{
preprocess_garmin_data(df_all = NULL, from = "", to = "", km = NULL)
}
\arguments{
\item{df_all}{data frame with all Garmin activity data}

\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}

\item{km}{numeric target distance in km for the period from from to to}
}
\value{
checked list of variables: from, to, km
}
\description{
Ensures that subsequent functions can be run on Garmin activity data.
}
\keyword{internal}
