% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndCalc.R
\name{process_data}
\alias{process_data}
\title{Load and process Garmin activity data}
\usage{
process_data(activityStr = "Running", fromStr = "", toStr = "", km = NULL)
}
\arguments{
\item{activityStr}{character string of activity type to use as a filter}

\item{fromStr}{character string of form "YYYY-MM-DD"}

\item{toStr}{character string of form "YYYY-MM-DD"}

\item{km}{numeric target distance in km for the period from fromStr to toStr}
}
\value{
saves processed data and plots to Output folder
}
\description{
Load and process Garmin activity data
}
