% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{process_load}
\alias{process_load}
\title{Load and process Garmin activity data to calculate training load}
\usage{
process_load(activityStr, fromStr, toStr)
}
\arguments{
\item{activityStr}{string of activity type to use as a filter}

\item{fromStr}{string of form "YYYY-MM-DD"}

\item{toStr}{string of form "YYYY-MM-DD"}
}
\value{
data frame with activities of type activityStr within date window
}
\description{
Load and process Garmin activity data to calculate training load
}
