% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressToTarget.R
\name{progress_to_target}
\alias{progress_to_target}
\title{Document progress towards target of Garmin activity data}
\usage{
progress_to_target(
  activity = "Running",
  df_all = NULL,
  from = "",
  to = "",
  km = NULL
)
}
\arguments{
\item{activity}{Type of activity (e.g., "Running") - only used for printing}

\item{df_all}{data frame with all Garmin activity data}

\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}

\item{km}{numeric target distance in km for the period from from to to}
}
\value{
saves processed data and plots to Output folder
}
\description{
Document progress towards target of Garmin activity data
}
