% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTreemap.R
\name{rename_km1}
\alias{rename_km1}
\title{Rename km1}
\usage{
rename_km1(x)
}
\arguments{
\item{x}{vector with 1 km distance bins}
}
\value{
character vector with renamed bins
}
\description{
Utility function to rename km1 bins for treemap plotting
}
\keyword{internal}
