% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTreemap.R
\name{rename_km5}
\alias{rename_km5}
\title{Rename km5}
\usage{
rename_km5(x)
}
\arguments{
\item{x}{vector with 5 km distance bins}
}
\value{
character vector with renamed bins
}
\description{
Utility function to rename km5 bins for treemap plotting
}
\keyword{internal}
