% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{sumDays}
\alias{sumDays}
\title{Utility function to sum load per day}
\usage{
sumDays(df, daydf)
}
\arguments{
\item{df}{data frame with Date and load columns}

\item{daydf}{data frame with all dates in the time window}
}
\value{
data frame with summed load per day
}
\description{
Utility function to sum load per day
}
\keyword{internal}
