% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseActivities.R
\name{summarise_activities}
\alias{summarise_activities}
\title{Wrapper to analyse and visualise Garmin activities}
\usage{
summarise_activities(
  activity = "Running",
  datadir = "Data",
  from = "",
  to = "",
  target = NULL,
  progress = TRUE,
  form = TRUE,
  tree = TRUE,
  calendar = TRUE
)
}
\arguments{
\item{activity}{Type of activity to filter (e.g., "Running") must be contained
in the 'Activity.Type' column of the Garmin data, e.g. Running will include
all runs, including treadmill runs.}

\item{datadir}{directory where Garmin CSV files are stored (default is "Data")}

\item{from}{start date in "YYYY-MM-DD" format}

\item{to}{end date in "YYYY-MM-DD" format}

\item{target}{optional numeric target distance in km for the period from
'from' to 'to'}

\item{progress}{logical indicating whether to generate progress towards target plot}

\item{form}{logical indicating whether to generate TSS form analysis plot}

\item{tree}{logical indicating whether to generate distance treemap plot}

\item{calendar}{logical indicating whether to generate calendar view plot}
}
\description{
This function loads Garmin activity data, filters it by activity type and date range,
and then generates several visualisations and analyses including progress towards a target,
Training Stress Score (TSS) form analysis, a distance treemap, and a
calendar view of activities.
}
