% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeatmap.R
\name{turbo}
\alias{turbo}
\title{Create turbo palette (taken from colorTK)}
\usage{
turbo(
  paletteLength.num = NULL,
  space.chr = "rgb",
  interpolate.chr = "linear",
  bias.num = 1
)
}
\arguments{
\item{paletteLength.num}{number of colors to generate.}

\item{space.chr}{character string; interpolation in RGB or CIE Lab color
spaces. See ?grDevices::colorRamp for more details. (Default "rgb")}

\item{interpolate.chr}{use spline or linear interpolation. See
?grDevices::colorRamp for more details. (Default "linear")}

\item{bias.num}{\if{html}{\out{<numeric>}}: a positive number.  Higher values give more widely
spaced colors at the high end. See ?grDevices::colorRamp for more details.
(Default 1)}
}
\value{
A vector of color.
}
\description{
Turbo is a jet alternative developed by Anton Mikhailov
}
\keyword{internal}
