% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressToTarget.R
\name{weekly_summary}
\alias{weekly_summary}
\title{Calculate weekly summary of distances}
\usage{
weekly_summary(df, from, to)
}
\arguments{
\item{df}{data frame with Date and Distance columns}

\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}
}
\value{
data frame with weekly total and daily average distances
}
\description{
Calculate weekly summary of distances
}
\keyword{internal}
