% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTreemap.R
\name{distanceTreemap}
\alias{distanceTreemap}
\title{Distance Treemap}
\usage{
distanceTreemap(df_all, from, to)
}
\arguments{
\item{df_all}{data frame with all Garmin activity data}

\item{from}{string of form "YYYY-MM-DD"}

\item{to}{string of form "YYYY-MM-DD"}
}
\description{
Make a treemap of run distances
}
