% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{find_form}
\alias{find_form}
\title{Function to find and plot Form (CTL, ATL, TSS)}
\usage{
find_form(df_all, from, to)
}
\arguments{
\item{df_all}{data frame with all Garmin activity data}

\item{from}{character string of form "YYYY-MM-DD"}

\item{to}{character string of form "YYYY-MM-DD"}
}
\value{
saves plot to Output/Plots folder
}
\description{
Function to find and plot Form (CTL, ATL, TSS)
}
