% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGarminData.R
\name{load_garmin_data}
\alias{load_garmin_data}
\title{Load all Garmin activity data from CSV files}
\usage{
load_garmin_data(activity = "Running", datadir = "Data", minimal = TRUE)
}
\arguments{
\item{activity}{character string of activity type to use as a filter}

\item{datadir}{character string of folder name where CSV files are stored}

\item{minimal}{logical indicating whether to load minimal columns (default TRUE)}
}
\value{
data frame of activity data
}
\description{
This function reads all Garmin activity CSV files from the specified data
directory.
}
